chore: move mockData and edit imports in filterContests feature
