+++ src/features/filterContests/index.ts
@@ -1,4 +1,20 @@
+import { selectFilters } from './model/selectors'
+import { filterActions, filterReducer } from './model/slice'
+import {
+    FilterData,
+    FilterItem,
+    FilterObject,
+    FilterSchema,
+} from './model/types'
 import FilterController from './ui/filterController'
 import FilterModal from './ui/filterModal/filterModal'
 
-export { FilterController, FilterModal }
+export type { FilterData, FilterItem, FilterObject, FilterSchema }
+
+export {
+    filterActions,
+    FilterController,
+    FilterModal,
+    filterReducer,
+    selectFilters,
+}
